refactor(weather): Move weather repositories to data layer
- Relocated weather-related repositories to lib/data/repositories/weather/repositories/
- Adjusted import paths across services and view models
- Promotes separation of concerns and aligns with Clean Architecture structure
