+++ lib/data/repositories/weather/repositories/unified_weather_repository.dart
@@ -1,5 +1,5 @@
-import '../../../core/network/response/result.dart';
-import '../models/forecast.dart';
+import '../../../../core/network/response/result.dart';
+import '../../../../domain/models/weather/models/forecast.dart';
 import 'firestore_weather_repository.dart';
 import 'weather_repository.dart';
